Move language logic to ThemeProvider instead of INSIDE LanguageSelector
